we can have higher framerate

+++ peerconnection/peerconnection.go
@@ -107,7 +107,7 @@ func NewPeerConnectionThread(
 		Video: func(constraint *mediadevices.MediaTrackConstraints) {
 			constraint.Width = prop.Int(1920)
 			constraint.Height = prop.Int(1080)
-			constraint.FrameRate = prop.Float(60)
+			constraint.FrameRate = prop.Float(100)
 		},
 		Codec: codecselector,
 	})
